🪛(fix): structure
